시/분/초침과 테두리 두께가 시계의 크기에 따라 변경되도록 수정

+++ domain/src/main/java/com/example/domain/models/CoordinateClockPartData.kt
@@ -9,6 +9,6 @@ data class CoordinateClockPartData(
     val secondColor : String,
     val uiStateData: UiStateData,
     val strokeColor : String,
-    val strokeWidth : Int,
+    val strokeWidth : Float,
     val useMiddleLineStroke : Boolean
 )
